+++ src/lib.rs
@@ -5,4 +5,5 @@ pub mod io;
 pub mod p3d;
 pub mod pbo;
 pub mod preprocess;
+pub mod run;
 pub mod sign;
